Rename project from origins-generator-service to prov-extractor
The scope of Origins has changed significantly since it’s inception.
The relevant change made to Origins for this repository is that it no
longer uses the PROV format natively for representation and storage.

That being said, the provenance being extracted from the supported
sources is still valid and useful in it’s own right. There is plenty
of tooling for W3C PROV formats.

Origins can still supports all the structures in a PROV document and
for that reason, the `origins` attributes will rename in the generated
output.

Signed-off-by: Byron Ruth <b@devel.io>